Separated test function from library function, made header file.
Main is now the only function in lbuxton-simp_test.c, and there is no
main in lbuxton-simp.c. There is now a header file lbuxton-simp.h
with the function prototypes for the library functions. The typedef
vstatus is also in the header and removed from lbuxton-simp.c.
This header in included in lbuxton-simp.c and lbuxton-simp_test.c.